LPS-71173 Move existing unit test and add a new test to check the valid case
